add main channel
